Split Android Gradle project into one library project and one demo application
May want to think more about package/folder names and where assets should live
