+++ android/build.gradle
@@ -7,31 +7,25 @@ buildscript {
     }
 }
 
-apply plugin: 'android'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    compile fileTree(dir: 'libs', include: '*.jar')
-    compile 'com.squareup.okhttp:okhttp:2.2.0'
-}
-
-android {
-
-    compileSdkVersion 19
-    buildToolsVersion "21.1.2"
-
-    sourceSets.main {
-        
-        manifest.srcFile 'AndroidManifest.xml'
-        java.srcDirs = ['src']
-        jni.srcDirs = []
-        jniLibs.srcDirs = ['libs']
-        resources.srcDirs = ['src']
-        res.srcDirs = ['res']
-        assets.srcDirs = ['../core/resources']
-
+allprojects {
+    repositories {
+        mavenCentral()
     }
+}
+
+project(":demo") {
+    apply plugin: 'com.android.application'
+
+    dependencies {
+        compile project(":tangram")
+    }
+}
+
+project(":tangram") {
+    apply plugin: 'com.android.library'
+
+    dependencies {
+        compile 'com.squareup.okhttp:okhttp:2.2.0'
+    }
+
 }
