+++ android/demo/build.gradle
@@ -1,0 +1,14 @@
+android {
+
+    compileSdkVersion 19
+    buildToolsVersion "21.1.2"
+
+    sourceSets.main {
+
+        manifest.srcFile 'AndroidManifest.xml'
+        java.srcDirs = ['src']
+        resources.srcDirs = ['src']
+        res.srcDirs = ['res']
+
+    }
+}
