+++ android/demo/res/drawable-ldpi/ic_launcher.png
