+++ android/demo/res/drawable-xhdpi/ic_launcher.png
