+++ android/demo/res/values/strings.xml
