+++ android/tangram/build.gradle
@@ -1,0 +1,15 @@
+android {
+
+    compileSdkVersion 19
+    buildToolsVersion "21.1.2"
+
+    sourceSets.main {
+
+        manifest.srcFile 'AndroidManifest.xml'
+        java.srcDirs = ['src']
+        jni.srcDirs = []
+        jniLibs.srcDirs = ['libs']
+        assets.srcDirs = ['../../core/resources']
+
+    }
+}
