+++ android/tangram/jni/platform_android.cpp
