+++ android/tangram/src/com/almeros/android/multitouch/BaseGestureDetector.java
