+++ android/tangram/src/com/almeros/android/multitouch/MoveGestureDetector.java
