+++ android/tangram/src/com/almeros/android/multitouch/RotateGestureDetector.java
