+++ android/tangram/src/com/almeros/android/multitouch/ShoveGestureDetector.java
