+++ android/tangram/src/com/almeros/android/multitouch/TwoFingerGestureDetector.java
